+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.5.0 - YYYY-MM-DD #
 
-- Plugin "git": show Git objects via `git show`.
+- Plugin "git": show Git revisions via `git show`.
 
 # 0.4.0 - 2016-03-04 #